Create a custom coloured slog.Logger and place in own package
