+++ internal/cmd/cmd.go
@@ -4,59 +4,16 @@ import (
 	"context"
 	"github.com/danielmichaels/doublestag/internal/config"
 	"github.com/danielmichaels/doublestag/internal/jobs"
+	"github.com/danielmichaels/doublestag/internal/logging"
 	"github.com/danielmichaels/doublestag/internal/store"
-	"github.com/danielmichaels/doublestag/internal/tracing"
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/riverqueue/river"
 	"io"
 	"log/slog"
-	"os"
 )
 
 type Globals struct{}
-
-type SlogHandler struct{ slog.Handler }
-
-func (s *SlogHandler) Handle(ctx context.Context, r slog.Record) error {
-	if traceID, ok := ctx.Value(tracing.TraceCtxKey).(string); ok {
-		r.Add("trace_id", slog.StringValue(traceID))
-	}
-	return s.Handler.Handle(ctx, r)
-}
-
-func (s *SlogHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
-	clone := *s
-	return &clone
-}
-
-func (s *SlogHandler) WithGroup(name string) slog.Handler {
-	return &SlogHandler{Handler: s.Handler.WithGroup(name)}
-}
-
-func createHandler(writer io.Writer, cfg *config.Conf) slog.Handler {
-	if cfg.AppConf.LogJson {
-		return slog.NewJSONHandler(writer, &slog.HandlerOptions{
-			Level: cfg.AppConf.LogLevel,
-		})
-	}
-	return slog.NewTextHandler(writer, &slog.HandlerOptions{
-		Level: cfg.AppConf.LogLevel,
-	})
-}
-
-func setupLogger(service string, cfg *config.Conf) (*slog.Logger, context.Context) {
-	var handler slog.Handler
-	handler = createHandler(os.Stderr, cfg)
-
-	handler = &SlogHandler{Handler: handler}
-
-	logger := slog.New(handler)
-	// default to none, will be overridden later
-	traceID := ""
-	ctx := context.WithValue(context.Background(), tracing.TraceCtxKey, traceID)
-	return logger, ctx
-}
 
 type Setup struct {
 	Config *config.Conf
@@ -105,7 +62,7 @@ func WithSilentLogging() SetupOption {
 
 func NewSetup(service string, opts ...SetupOption) (*Setup, error) {
 	cfg := config.AppConfig()
-	logger, lctx := setupLogger(service, cfg)
+	logger, lctx := logging.SetupLogger(service, cfg)
 	ctx, cancel := context.WithCancel(lctx)
 
 	db, err := store.NewDatabasePool(ctx, cfg)
